phase_02/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
psycopg2-binary==2.9.9
alembic==1.13.1
uuid-utils==0.6.0
better-exceptions==0.3.2
cryptography==41.0.7
bcrypt==4.1.2